Extract AgendaActions component
